Update highlight.js
This updates highlight.js to 10.7.3, which picks up various syntax highlighting fixes.
This is keeping on the 10.x release due to #1622.

I have also added a few requested languages: abnf

Note that there is a new deprecation in this release. I haven't updated the javascript code to avoid breaking any projects which just override highlight.js in their theme. Unfortunately this will result in a deprecation warning in the console, which can't easily be turned off.

Closes #1344
Closes #1674